+++ assets/javascripts/kitten/components/steppers/vertical-progress/index.js
@@ -12,7 +12,6 @@ const StyledContainer = styled.div`
     border: ${pxToRem(2)} solid ${COLORS.line1};
     padding: ${pxToRem(15)};
     border-radius: ${pxToRem(8)};
-    max-width: ${pxToRem(400)};
 
     @media (min-width: ${ScreenConfig.S.min}px) {
       padding: ${pxToRem(30)};