+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: Use `gap` on Flex containers whenever feasible.
 
+## [3.25.1] - 2021-07-09
+
+Fix:
+
+- `VerticalProgress`: Remove `max-width` on `k-VerticalProgress__wrapper`.
+
 ## [3.25.0] - 2021-07-08
 
 Feature: